+++ pkg/util/http.go
@@ -18,20 +18,13 @@ package util
 
 import (
 	"crypto/tls"
-	"fmt"
 	"net/http"
-	"os"
 	"time"
-
-	"github.com/go-logr/logr"
-	ctrl "sigs.k8s.io/controller-runtime"
 )
 
 var disableKeepAlives bool
-var minTLSVersion uint16
 
 func init() {
-	setupLog := ctrl.Log.WithName("http_setup")
 	var err error
 	// This code will be removed in https://github.com/kedacore/keda/pull/4191
 	// nosemgrep: trailofbits.go.invalid-usage-of-modified-variable.invalid-usage-of-modified-variable
@@ -39,36 +32,11 @@ func init() {
 	if err != nil {
 		disableKeepAlives = false
 	}
-
-	minTLSVersion = initMinTLSVersion(setupLog)
-}
-
-func initMinTLSVersion(logger logr.Logger) uint16 {
-	version, found := os.LookupEnv("KEDA_HTTP_MIN_TLS_VERSION")
-	minVersion := tls.VersionTLS12
-	if found {
-		switch version {
-		case "TLS13":
-			minVersion = tls.VersionTLS13
-		case "TLS12":
-			minVersion = tls.VersionTLS12
-		case "TLS11":
-			minVersion = tls.VersionTLS11
-		case "TLS10":
-			minVersion = tls.VersionTLS10
-		default:
-			logger.Info(fmt.Sprintf("%s is not a valid value, using `TLS12`. Allowed values are: `TLS13`,`TLS12`,`TLS11`,`TLS10`", version))
-			minVersion = tls.VersionTLS12
-		}
-	}
-	return uint16(minVersion)
 }
 
 func init() {
-	setupLog := ctrl.Log.WithName("http_setup")
 	disableKeepAlives = getKeepAliveValue()
 	rootCAs = getRootCAs()
-	minTLSVersion = initMinTLSVersion(setupLog)
 }
 
 func getKeepAliveValue() bool {
@@ -120,8 +88,4 @@ func CreateHTTPTransportWithTLSConfig(config *tls.Config) *http.Transport {
 		transport.IdleConnTimeout = 100 * time.Second
 	}
 	return transport
-}
-
-func GetMinTLSVersion() uint16 {
-	return minTLSVersion
 }
