+++ pkg/util/tls_config.go
@@ -21,31 +21,18 @@ import (
 	"crypto/x509"
 	"encoding/pem"
 	"fmt"
+	"os"
 
+	"github.com/go-logr/logr"
 	"github.com/youmark/pkcs8"
+	ctrl "sigs.k8s.io/controller-runtime"
 )
 
-func decryptClientKey(clientKey, clientKeyPassword string) ([]byte, error) {
-	block, _ := pem.Decode([]byte(clientKey))
+var minTLSVersion uint16
 
-	key, err := pkcs8.ParsePKCS8PrivateKey(block.Bytes, []byte(clientKeyPassword))
-	if err != nil {
-		return nil, err
-	}
-
-	pemData, err := x509.MarshalPKCS8PrivateKey(key)
-	if err != nil {
-		return nil, err
-	}
-
-	var pemPrivateBlock = &pem.Block{
-		Type:  "RSA PRIVATE KEY",
-		Bytes: pemData,
-	}
-
-	encodedData := pem.EncodeToMemory(pemPrivateBlock)
-
-	return encodedData, nil
+func init() {
+	setupLog := ctrl.Log.WithName("tls_setup")
+	minTLSVersion = initMinTLSVersion(setupLog)
 }
 
 // NewTLSConfigWithPassword returns a *tls.Config using the given ceClient cert, ceClient key,
@@ -84,11 +71,61 @@ func NewTLSConfig(clientCert, clientKey, caCert string, unsafeSsl bool) (*tls.Co
 	return NewTLSConfigWithPassword(clientCert, clientKey, "", caCert, unsafeSsl)
 }
 
-// createTLSClientConfig returns a new TLS Config
+// CreateTLSClientConfig returns a new TLS Config
 // unsafeSsl parameter allows to avoid tls cert validation if it's required
 func CreateTLSClientConfig(unsafeSsl bool) *tls.Config {
 	return &tls.Config{
 		InsecureSkipVerify: unsafeSsl,
 		RootCAs:            getRootCAs(),
+		MinVersion:         GetMinTLSVersion(),
 	}
+}
+
+// GetMinTLSVersion return the minTLSVersion based on configurations
+func GetMinTLSVersion() uint16 {
+	return minTLSVersion
+}
+
+func initMinTLSVersion(logger logr.Logger) uint16 {
+	version, found := os.LookupEnv("KEDA_HTTP_MIN_TLS_VERSION")
+	minVersion := tls.VersionTLS12
+	if found {
+		switch version {
+		case "TLS13":
+			minVersion = tls.VersionTLS13
+		case "TLS12":
+			minVersion = tls.VersionTLS12
+		case "TLS11":
+			minVersion = tls.VersionTLS11
+		case "TLS10":
+			minVersion = tls.VersionTLS10
+		default:
+			logger.Info(fmt.Sprintf("%s is not a valid value, using `TLS12`. Allowed values are: `TLS13`,`TLS12`,`TLS11`,`TLS10`", version))
+			minVersion = tls.VersionTLS12
+		}
+	}
+	return uint16(minVersion)
+}
+
+func decryptClientKey(clientKey, clientKeyPassword string) ([]byte, error) {
+	block, _ := pem.Decode([]byte(clientKey))
+
+	key, err := pkcs8.ParsePKCS8PrivateKey(block.Bytes, []byte(clientKeyPassword))
+	if err != nil {
+		return nil, err
+	}
+
+	pemData, err := x509.MarshalPKCS8PrivateKey(key)
+	if err != nil {
+		return nil, err
+	}
+
+	var pemPrivateBlock = &pem.Block{
+		Type:  "RSA PRIVATE KEY",
+		Bytes: pemData,
+	}
+
+	encodedData := pem.EncodeToMemory(pemPrivateBlock)
+
+	return encodedData, nil
 }
